Move inventory to side
