+++ client/src/components/inventory/Inventory.tsx
@@ -1,21 +1,19 @@
 import React from 'react';
 import styled from 'styled-components';
-import { Box } from '@material-ui/core';
-import InventoryRow from './InventoryRow';
+import { useDrop } from 'react-dnd';
 import { CourseData, ClassData } from '../../interfaces/CourseData';
 
-const StyledInventory = styled(Box)`
-// display: none;
-flex-direction: column;
+import InventoryCourseClass from './InventoryCourseClass';
 
-border: 1px solid;
-box-sizing: border-box;
-margin-top: 60px;
-margin-bottom: 30px;
-border-color: ${(props) => props.theme.palette.secondary.main};
+const StyledInventory = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 2px;
+  grid-column: -2;
+  grid-row: 2 / -1;
 `;
 
-interface InventoryProps {
+export interface InventoryProps {
   selectedCourses: CourseData[]
   selectedClasses: ClassData[]
   assignedColors: Record<string, string>
@@ -27,20 +25,52 @@ const Inventory: React.FC<InventoryProps> = ({
   selectedClasses,
   assignedColors,
   removeClass,
-}) => (
-  <div>
-    <StyledInventory>
-      {selectedCourses.length
-        ? (
-          <InventoryRow
-            selectedCourses={selectedCourses}
-            selectedClasses={selectedClasses}
-            removeClass={removeClass}
-          />
-        )
-        : 'No courses have been selected'}
+}) => {
+  const getInventoryCourseClasses = (): React.ReactNode[] => {
+    let classNodes: React.ReactNode[] = []
+
+    // return course classes for activities which don't currently have a selected class
+    selectedCourses.forEach((course) => {
+      Object.entries(course.activities).forEach(([activity, activityClasses]) => {
+        if (!activityClasses.some(
+            (classData) => selectedClasses.includes(classData),
+          )
+        ) {
+          classNodes.push(
+            <InventoryCourseClass
+              key={`${course.code}-${activity}`}
+              courseCode={course.code}
+              activity={activity}
+              color={assignedColors[course.code]}
+            />
+          )
+        }
+      })
+    })
+
+    return classNodes;
+  };
+
+  const ids = selectedCourses.reduce<string[]>((array, course): string[] => (
+    [...array, ...Object.keys(course.activities).map((activity) => (
+      `${course.code}-${activity}`
+    ))]
+  ), []);
+
+  const [{ canDrop }, drop] = useDrop({
+    accept: ids,
+    drop: ({ classData }: any) => removeClass(classData),
+    collect: (monitor) => ({
+      canDrop: monitor.canDrop(),
+      isOver: monitor.isOver(),
+    }),
+  });
+
+  return (
+    <StyledInventory ref={drop}>
+      {getInventoryCourseClasses()}
     </StyledInventory>
-  </div>
-);
+  );
+};
 
 export default Inventory;
